move test utilities to dir
